webservice/requirements.txt
Django==1.5.2
South==0.8.2
django-filter
django-reversion
django-suit
django-mptt==0.6.0
django-extensions
django-model-utils
Pillow>=2.1.0
djrill
markdown
html5lib
html2text
sh
six
pytz
django-sizefield
#-e git+git://github.com/Soaa-/django-nested-inlines.git#egg=django-nested-inlines
django-rest-swagger
djangorestframework==2.3.7
drf-extensions
#django-userena
django-comments-xtd==1.2a1

# django-admin-user-status dependents
#django-qsstats-magic>=0.6.1
python-dateutil
#django-chart-tools>=0.2.1
#django-admin-user-stats
pysolr
django-haystack

django-redis==3.3
django-redis-sessions==0.4.0

mongoengine

# frontend
django-static-precompiler
django-compressor

# haml templatetags
https://github.com/killuavx/HamlPy/archive/support/py3.zip#egg=HamlPy
djaml

https://github.com/killuavx/django-widgets/archive/support/py3.zip#egg=django-widgets
feedparser
Mezzanine==3.0.8

# templatetags
https://github.com/killuavx/django-template-utils/archive/support/py3.zip#egg=django-template-utils

qrcode
drest

#django-floppyforms
#django-daterange-filter
django-chart-tools
django-classy-tags
django-meta
django-meta-mixin


# django-memoize
https://github.com/akanouras/django-memoize/archive/akanouras/old-django-compat.zip
django-import-export
# 2to3
django-url-tools

celery[redis]
django-celery
flower
pyyaml
ua-parser
user-agents
django-user-agents

https://github.com/killuavx/cache-tagging/archive/support/py3.zip

jieba3k

https://github.com/senko/python-video-converter/archive/master.zip#egg=python-video-coverter
